Disabling package removal
After a lot of reports, I'm disabling the package removal until a better solution is find. The problem is: we need to find a way to ask the user what should be removed and what shouldn't during the uninstall process, because part of the packages probably are in use by other peoples or softwares.